refactor: 合併 StatisticsQueryService 至 StatisticsApplicationService
- 移除 StatisticsQueryServiceInterface（只有一個實作，過度設計）
- 移除 StatisticsQueryService，25 個方法全數搬至 StatisticsApplicationService
- 更新 4 個呼叫端（StatisticsController、StatisticsAdminController、
  StatisticsExportService、StatisticsQueryAdapter）
- 更新 StatisticsServiceProvider 移除介面綁定
- 更新對應測試
