+++ backend/app/Infrastructure/Statistics/Services/StatisticsExportService.php
@@ -4,22 +4,22 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Statistics\Services;
 
+use App\Application\Services\Statistics\DTOs\StatisticsQueryDTO;
+use App\Application\Services\Statistics\StatisticsApplicationService;
 use App\Domains\Statistics\Contracts\BatchExportResult;
 use App\Domains\Statistics\Contracts\ExportResult;
 use App\Domains\Statistics\Contracts\StatisticsExportServiceInterface;
 use App\Domains\Statistics\Contracts\StatisticsFormatterInterface;
-use App\Domains\Statistics\Contracts\StatisticsQueryServiceInterface;
 use DateTime;
+use DateTimeImmutable;
 use InvalidArgumentException;
 use RuntimeException;
 use Throwable;
 
 final class StatisticsExportService implements StatisticsExportServiceInterface
 {
-    /** 預設匯出格式 */
     private const DEFAULT_FORMAT = 'json';
 
-    /** 支援的統計類型 */
     private const SUPPORTED_TYPES = [
         'overview',
         'posts',
@@ -32,7 +32,7 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
      * @param array<string, StatisticsFormatterInterface> $formatters 格式化器陣列
      */
     public function __construct(
-        private readonly StatisticsQueryServiceInterface $queryService,
+        private readonly StatisticsApplicationService $queryService,
         private readonly array $formatters,
     ) {}
 
@@ -43,16 +43,13 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         $this->validateFormat($format);
 
         try {
-            // 準備查詢選項
-            $queryOptions = $this->prepareQueryOptions($options);
-            // 取得統計資料
-            /** @phpstan-ignore-next-line argument.type */
-            $data = $this->queryService->getOverview($queryOptions);
-            // 格式化資料
+            $queryDTO = $this->prepareQueryDTO($options);
+            $overview = $this->queryService->getOverview($queryDTO);
+            /** @var array<string, mixed> $data */
+            $data = $overview->toArray();
             $formatter = $this->formatters[$format];
             $content = $formatter->format($data, $options);
 
-            // 建立匯出結果
             return new ExportResult(
                 format: $format,
                 filename: $formatter->getRecommendedFilename('overview', $options),
@@ -74,9 +71,10 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         $this->validateFormat($format);
 
         try {
-            $queryOptions = $this->prepareQueryOptions($options);
-            /** @phpstan-ignore-next-line argument.type */
-            $data = $this->queryService->getPostStatistics($queryOptions);
+            $queryDTO = $this->prepareQueryDTO($options);
+            $paginated = $this->queryService->getPostStatistics($queryDTO);
+            /** @var array<string, mixed> $data */
+            $data = $paginated->toArray();
             $formatter = $this->formatters[$format];
             $content = $formatter->format($data, $options);
 
@@ -101,9 +99,9 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         $this->validateFormat($format);
 
         try {
-            $queryOptions = $this->prepareQueryOptions($options);
-            /** @phpstan-ignore-next-line argument.type */
-            $data = $this->queryService->getSourceDistribution($queryOptions);
+            $queryDTO = $this->prepareQueryDTO($options);
+            /** @var array<string, mixed> $data */
+            $data = $this->queryService->getSourceDistribution($queryDTO);
             $formatter = $this->formatters[$format];
             $content = $formatter->format($data, $options);
 
@@ -128,9 +126,10 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         $this->validateFormat($format);
 
         try {
-            $queryOptions = $this->prepareQueryOptions($options);
-            /** @phpstan-ignore-next-line argument.type */
-            $data = $this->queryService->getUserStatistics($queryOptions);
+            $queryDTO = $this->prepareQueryDTO($options);
+            $paginated = $this->queryService->getUserStatistics($queryDTO);
+            /** @var array<string, mixed> $data */
+            $data = $paginated->toArray();
             $formatter = $this->formatters[$format];
             $content = $formatter->format($data, $options);
 
@@ -155,9 +154,9 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         $this->validateFormat($format);
 
         try {
-            $queryOptions = $this->prepareQueryOptions($options);
-            /** @phpstan-ignore-next-line argument.type */
-            $data = $this->queryService->getPopularContent($queryOptions);
+            $queryDTO = $this->prepareQueryDTO($options);
+            /** @var array<string, mixed> $data */
+            $data = $this->queryService->getPopularContent($queryDTO);
             $formatter = $this->formatters[$format];
             $content = $formatter->format($data, $options);
 
@@ -227,9 +226,6 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         return self::SUPPORTED_TYPES;
     }
 
-    /**
-     * 驗證匯出格式是否支援.
-     */
     private function validateFormat(string $format): void
     {
         if (!isset($this->formatters[$format])) {
@@ -237,9 +233,6 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         }
     }
 
-    /**
-     * 驗證統計類型是否支援.
-     */
     private function validateStatisticsType(string $type): void
     {
         if (!in_array($type, self::SUPPORTED_TYPES, true)) {
@@ -248,44 +241,42 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
     }
 
     /**
-     * 準備查詢選項.
-     *
-     * @return array<string, mixed>
+     * 將匯出選項轉換為 StatisticsQueryDTO.
      */
-    private function prepareQueryOptions(array $options): array
+    private function prepareQueryDTO(array $options): StatisticsQueryDTO
     {
-        $queryOptions = [];
-        // 處理時間範圍
-        if (isset($options['period_start'])) {
-            $queryOptions['period_start'] = $options['period_start'];
-        }
-        if (isset($options['period_end'])) {
-            $queryOptions['period_end'] = $options['period_end'];
-        }
-        // 處理分頁
-        if (isset($options['limit'])) {
-            $queryOptions['limit'] = $options['limit'];
-        }
-        if (isset($options['offset'])) {
-            $queryOptions['offset'] = $options['offset'];
-        }
-        // 處理其他選項
-        if (isset($options['include_details'])) {
-            $queryOptions['include_details'] = $options['include_details'];
-        }
-        if (isset($options['group_by_detail'])) {
-            $queryOptions['group_by_detail'] = $options['group_by_detail'];
-        }
-        if (isset($options['include_inactive'])) {
-            $queryOptions['include_inactive'] = $options['include_inactive'];
+        $startDate = null;
+        $endDate = null;
+
+        if (isset($options['period_start']) && $options['period_start'] instanceof DateTimeImmutable) {
+            $startDate = $options['period_start'];
+        } elseif (isset($options['period_start']) && $options['period_start'] instanceof DateTime) {
+            $startDate = DateTimeImmutable::createFromMutable($options['period_start']);
         }
 
-        return $queryOptions;
+        if (isset($options['period_end']) && $options['period_end'] instanceof DateTimeImmutable) {
+            $endDate = $options['period_end'];
+        } elseif (isset($options['period_end']) && $options['period_end'] instanceof DateTime) {
+            $endDate = DateTimeImmutable::createFromMutable($options['period_end']);
+        }
+
+        $page = 1;
+        $limit = 20;
+        if (isset($options['limit']) && is_numeric($options['limit'])) {
+            $limit = min(100, max(1, (int) $options['limit']));
+        }
+        if (isset($options['offset']) && is_numeric($options['offset'])) {
+            $page = (int) floor((int) $options['offset'] / $limit) + 1;
+        }
+
+        return new StatisticsQueryDTO(
+            startDate: $startDate,
+            endDate: $endDate,
+            page: $page,
+            limit: $limit,
+        );
     }
 
-    /**
-     * 計算記錄數.
-     */
     private function countRecords(array $data): int
     {
         $count = 0;
@@ -304,17 +295,12 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
         return $count;
     }
 
-    /**
-     * 檢查是否為順序陣列.
-     */
     private function isSequentialArray(array $array): bool
     {
         return array_keys($array) === range(0, count($array) - 1);
     }
 
     /**
-     * 建立元資料.
-     *
      * @return array<string, mixed>
      */
     private function buildMetadata(array $options, array $data): array
@@ -323,14 +309,23 @@ final class StatisticsExportService implements StatisticsExportServiceInterface
             'export_time'    => new DateTime()->format('Y-m-d H:i:s'),
             'format_options' => array_diff_key($options, array_flip(['format'])),
         ];
-        // 添加資料相關的元資料
         if (isset($options['include_details'])) {
             $metadata['include_details'] = $options['include_details'];
         }
         if (isset($options['period_start'], $options['period_end'])) {
+            $start = $options['period_start'];
+            $end = $options['period_end'];
+            /** @var string $startStr */
+            $startStr = $start instanceof DateTime || $start instanceof DateTimeImmutable
+                ? $start->format('Y-m-d')
+                : (is_string($start) ? $start : '');
+            /** @var string $endStr */
+            $endStr = $end instanceof DateTime || $end instanceof DateTimeImmutable
+                ? $end->format('Y-m-d')
+                : (is_string($end) ? $end : '');
             $metadata['period'] = [
-                'start' => $options['period_start']->format('Y-m-d'),
-                'end'   => $options['period_end']->format('Y-m-d'),
+                'start' => $startStr,
+                'end'   => $endStr,
             ];
         }
